Version 1.7.6 New mod added: RDR2 Mix 2 New game added: The Chant New guides added: RDR2 Mix 2, High On Life, The Chant
Some Windows 11/Linux users are experiencing an error when running the .exe file. This issue will be addressed in upcoming updates. In the meantime, you can try resolving the error temporarily by installing the two fonts located in the Fonts folder alongside the .exe file. Simply open the files Segoe UI and CustomTkinter_shapes_font. This may temporarily resolve the issue.
It's also recommended to disable the antivirus when running the Utility, as it may cause conflicts with some mods.